backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# Auth & Security
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
cryptography==44.0.0

# Settings
pydantic[email]==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# Task queue
celery==5.4.0
redis==5.2.1
kombu==5.4.2

# HTTP client (for Questrade API)
httpx==0.28.1
aiohttp==3.11.10

# Data processing
pandas>=2.2.3
numpy>=2.0.0
pandas-ta>=0.4.67b0

# WebSockets
websockets==14.1

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
factory-boy==3.3.1
aiosqlite==0.20.0

# Development
black==24.10.0
ruff==0.8.4
